Fix arguments that don't match the parameter types
- in_array() with a null haystack when the "roles" request parameter
  is missing (users/add_group and users/group_types dashboard pages)
- in_array() with a Traversable haystack in SeoCanonical (the list of
  the query string parameters is now always stored as an array)
- random_bytes() called with a float
- setStartDateAllDay() called with integers instead of booleans
- UserInfo::getByID() called with a SimpleXMLElement
- t2() called with floats (rounded/floored values are now cast to int)
- Page::getByID() called with the string IDs of the bulk dialogs
- setCustomRequestUser() called with -1/false instead of null
- Key::setPermissionObject() called with false instead of null

+++ concrete/src/Api/OAuth/Client/ClientFactory.php
@@ -95,7 +95,7 @@ class ClientFactory
      */
     protected function generateString($length)
     {
-        $bytes = ceil($length / 2);
+        $bytes = (int) ceil($length / 2);
         $string = bin2hex(random_bytes($bytes));
 
         return substr($string, 0, $length);
